+++ src/libtcod/list.h
@@ -42,7 +42,10 @@ struct TCOD_List;
 
 typedef struct TCOD_List* TCOD_list_t;
 
-TCODLIB_API TCOD_list_t TCOD_list_new(void);
+TCODLIB_API
+TCOD_DEPRECATED("TCOD_List is not a suitable container.  Use a custom array or a C++ container instead.")
+TCOD_list_t TCOD_list_new(void);
+
 TCODLIB_API TCOD_list_t TCOD_list_allocate(int nb_elements);
 TCODLIB_API TCOD_list_t TCOD_list_duplicate(TCOD_list_t l);
 TCODLIB_API void TCOD_list_delete(TCOD_list_t l);